Extract getUsedAmpExtensions and add tests
